Add a function to assign a setting; this permits testing for emptiness/non-existence in a template.
It seems more natural to name the assignment tag `get_setting`, and the simple tag to output it as `setting`, so the new tag takes over the old name in this change.

A consequence of this is that the function to get the value (test setting type, return appropriate field) is now a method on the object; that it can now be used outside of templates without duplicating the same code is a nice side-effect.